icelandic-voice-agent/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
websockets>=14.0
anthropic>=0.45.0
google-cloud-speech>=2.29.0
azure-cognitiveservices-speech>=1.42.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
twilio>=9.4.0
scipy>=1.14.0
numpy>=2.1.0
python-dotenv>=1.0.1
structlog>=24.4.0
httpx>=0.28.0
audioop-lts>=0.2.1
grpcio>=1.68.0
